HealthCheck interface. Pricefeed daemon implementation of interface. No tests, no healthcheck updates on task failure.
Checkpoint - optional refactor of subtaskrunner into client methods.

Back out subtaskrunner changes.

Test cases for daemon health.

Channel cleanup.

Undo completeStartup and fix comment.

Comments.

mega-nit: missing period.

remove daemon client. update internal health indicator to track error on pricefeed daemon.

Pricefeed updates with new health checkable shared implementation.

Updates to health-checkable - test cases, specific error messages.

Migrate pricefeed changes. Lint fix for health check tests.

Remove accidental addition to this branch.

Remove explicit toggle to unhealthy state. This will be caught after 5 min.

Comment cleanup.

PR feedback.

Performance improvement - reduce test time from 100ms to sub-ms for health-check tests.

Refactor test util method since it will be used for all daemons.

Update protocol/daemons/types/health_checkable.go

Update protocol/daemons/types/health_checkable.go

Update protocol/daemons/types/health_checkable.go

Update protocol/daemons/types/health_checkable.go

Update protocol/daemons/types/health_checkable.go

Update protocol/daemons/types/health_checkable.go

Remove timestamp parameter from health check reporth methods.

Update comment for consistency.

Checkpoint.

Add sub task runner file.

partial progress through test.

Tests.

Update liquidations with rebase changes from previous PR.

Update with rebase.

Refactor - move implementation methods and tests into separate files.

Store liquidations client on app.

PR nits.
